+++ src/main/java/org/opentripplanner/street/model/edge/VehicleParkingEdge.java
@@ -1,6 +1,7 @@
 package org.opentripplanner.street.model.edge;
 
 import org.locationtech.jts.geom.LineString;
+import org.opentripplanner.framework.i18n.I18NString;
 import org.opentripplanner.routing.api.request.StreetMode;
 import org.opentripplanner.routing.api.request.preference.BikePreferences;
 import org.opentripplanner.routing.api.request.preference.CarPreferences;
@@ -10,7 +11,6 @@ import org.opentripplanner.street.model.vertex.VehicleParkingEntranceVertex;
 import org.opentripplanner.street.search.TraverseMode;
 import org.opentripplanner.street.search.state.State;
 import org.opentripplanner.street.search.state.StateEditor;
-import org.opentripplanner.transit.model.basic.I18NString;
 
 /**
  * Parking a vehicle edge.