+++ src/main/java/org/opentripplanner/street/model/vertex/StationElementVertex.java
@@ -1,8 +1,8 @@
 package org.opentripplanner.street.model.vertex;
 
 import javax.annotation.Nonnull;
+import org.opentripplanner.framework.i18n.I18NString;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.transit.model.basic.I18NString;
 import org.opentripplanner.transit.model.site.StationElement;
 
 public abstract class StationElementVertex extends Vertex {