+++ src/main/java/org/opentripplanner/transit/model/site/MultiModalStation.java
@@ -3,8 +3,8 @@ package org.opentripplanner.transit.model.site;
 import java.util.Collection;
 import java.util.Objects;
 import javax.annotation.Nonnull;
-import org.opentripplanner.transit.model.basic.I18NString;
-import org.opentripplanner.transit.model.basic.WgsCoordinate;
+import org.opentripplanner.framework.geometry.WgsCoordinate;
+import org.opentripplanner.framework.i18n.I18NString;
 import org.opentripplanner.transit.model.framework.AbstractTransitEntity;
 import org.opentripplanner.transit.model.framework.FeedScopedId;