+++ src/test/java/org/opentripplanner/street/integration/CarRoutingTest.java
@@ -13,7 +13,7 @@ import org.locationtech.jts.geom.Geometry;
 import org.opentripplanner.ConstantsForTests;
 import org.opentripplanner.TestOtpModel;
 import org.opentripplanner._support.time.ZoneIds;
-import org.opentripplanner.framework.geometry.PolylineEncoder;
+import org.opentripplanner.framework.geometry.EncodedPolyline;
 import org.opentripplanner.model.GenericLocation;
 import org.opentripplanner.model.plan.StreetLeg;
 import org.opentripplanner.routing.algorithm.mapping.GraphPathToItineraryMapper;
@@ -159,6 +159,6 @@ public class CarRoutingTest {
         })
     );
     Geometry legGeometry = itineraries.get(0).getLegs().get(0).getLegGeometry();
-    return PolylineEncoder.encodeGeometry(legGeometry).points();
+    return EncodedPolyline.encode(legGeometry).points();
   }
 }