+++ docs/Changelog.md
@@ -31,6 +31,11 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - fix: forBoarding and forAlighting on cancelled calls in transmodel api [#4639](https://github.com/opentripplanner/OpenTripPlanner/pull/4639)
 - Add test for keeping Transmodel GraphQL schema in sync [#4658](https://github.com/opentripplanner/OpenTripPlanner/pull/4658)
 - Make maxFlexTripDuration configurable [#4642](https://github.com/opentripplanner/OpenTripPlanner/pull/4642)
+- Extract separate conditions for , ,  during OSM tag mapping, add Portland tag mapping [#4593](https://github.com/opentripplanner/OpenTripPlanner/pull/4593)
+- Better fallback in shape geometry processing [#4617](https://github.com/opentripplanner/OpenTripPlanner/pull/4617)
+- Cleanup org.opentripplanner.openstreetmap package structure [#4661](https://github.com/opentripplanner/OpenTripPlanner/pull/4661)
+- Fix batch queries in legacy graphql API [#4663](https://github.com/opentripplanner/OpenTripPlanner/pull/4663)
+- Add filter for minimum bike parking distance [#4626](https://github.com/opentripplanner/OpenTripPlanner/pull/4626)
 [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)